+++ frontend/common/src/components/Link/index.ts
@@ -2,11 +2,19 @@ import DownloadCsv from "./DownloadCsv";
 import ExternalLink from "./ExternalLink";
 import Link from "./Link";
 import IconLink from "./IconLink";
+import ScrollToLink from "./ScrollToLink";
 import type { ExternalLinkProps } from "./ExternalLink";
 import type { LinkProps } from "./Link";
 import type { IconLinkProps } from "./IconLink";
 import type { DownloadCsvProps } from "./DownloadCsv";
+import type { ScrollToLinkProps } from "./ScrollToLink";
 
 export default Link;
-export { DownloadCsv, ExternalLink, IconLink };
-export type { DownloadCsvProps, ExternalLinkProps, LinkProps, IconLinkProps };
+export { DownloadCsv, ExternalLink, IconLink, ScrollToLink };
+export type {
+  DownloadCsvProps,
+  ExternalLinkProps,
+  LinkProps,
+  IconLinkProps,
+  ScrollToLinkProps,
+};
